+++ assets/styles.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  header: {
+    fontSize: 25,
+  },
+  image: {
+    marginTop: 10,
+    width: 150,
+    height: 150,
+    borderColor: 'rgba(0,0,0,0.2)',
+    borderWidth: 3,
+    borderRadius: 100,
+  },
+});
+
+export default styles;
